requirements.txt
selenium
asgiref==3.2.3
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
Click==7.0
decorator==4.4.2
Django==2.2.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-leaflet==0.25.0
django-progressive-web-app==0.1.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
future==0.18.2
geocoder==1.38.1
idna==2.8
Pillow==6.2.1
postgis==1.0.4
progressbar==2.5
psycopg2==2.8.4
pytz==2019.3
ratelim==0.1.6
requests==2.23.0
six==1.13.0
South==1.0.2
sqlparse==0.3.0
urllib3==1.25.8
uwsgi==2.0.18
automium-web==0.1.1
beautifulsoup4==4.6.0
blinker==1.4
cffi==1.13.2
cryptography==2.1.4
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
legacy==0.1.6
lxml==4.4.1
MarkupSafe==1.1.1
pycparser==2.19
Werkzeug==0.16.0